+++ eslint.config.mjs
@@ -6,6 +6,7 @@ const eslintConfig = [
   {
     ignores: [
       "**/.next/**",
+      "**/out/**",
       "**/next-env.d.ts",
       "**/node_modules/**",
       "**/dist/**",